Переезд на hibernate Исправлена конфигурация DAOConfig Изменен employee контроллер Исправлены сервисы. Переделаны dao

+++ src/main/java/ru/technoserv/dao/EmployeeRowMapper.java
@@ -9,7 +9,7 @@ public class EmployeeRowMapper implements RowMapper<Employee> {
 
     public Employee mapRow(ResultSet resultSet, int i) throws SQLException {
         Employee employee = new Employee();
-        employee.setEmpID((Integer) resultSet.getInt("EMP_ID"));
+        employee.setEmpID(resultSet.getInt("EMP_ID"));
         //employee.setPosition((String) resultSet.getObject("TITLE"));
        // employee.setGrade((String) resultSet.getObject("DESCRIPTION"));
       //  employee.setDepartment((String) resultSet.getObject("DEPT_NAME"));
